Extract shared code between Loader and Saver
The Saver needs similar methods to the Loader for the answers. This
code has been extracted so it can be used by both classes.

Implements validation on saving node and group answers
